Fix tests in windows
